Initial template as servlet
This is the initial template to demonstrate that
we can react correctly to a REST call when used by Gerrit.
This is simply answering to $GERRITURL/plugins/gerrit-support/
writing something and a List in the response.
The List part is explicitly using scala runtime via List to be sure
we are using correctly it from gerrit.
The sbt or bazel packager must embed needed runtime in final .jar

Change-Id: I75c6ce69cf07bcc5bc75482e3e2f20ca415d869a